feat: implement Assembler::add_modules_from_dir()
